pyproject.toml
[tool.poetry]
name = "api-consumer"
package-mode = false
description = ""
authors = ["Danil Mikhailenko <danil.mikhailenko@t-systems.com>"]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"


[tool.ruff]
line-length = 120
src = ["app"]
target-version = "py311"

[tool.ruff.lint]
# See https://beta.ruff.rs/docs/rules/ for a full list of rules.
select = [
    "E", "W", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "PL", # pylint
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
